fix: 파일 조회 CodeRabbit 피드백 반영 (#355)
- offset int overflow 수정(page*size → (long) 캐스팅, CompanyFileCriteria.offset long화) + 경계 테스트
- vc 완료버전 집계를 회사 스코프로 먼저 좁힘 — 전체 file_version 스캔 방지(성능)
- MyFileController 설명을 실제 정책에 맞춤(전역 ADMIN/MASTER 는 우회가 아니라 스텝 정책상 EDITOR)

+++ src/main/java/com/group3/vitamins/file/application/query/CompanyFileCriteria.java
@@ -7,7 +7,7 @@ package com.group3.vitamins.file.application.query;
  * @param keyword    파일명·원본명·업로더 검색어 (nullable)
  * @param projectId  프로젝트 필터 (nullable)
  * @param extension  확장자 필터 (nullable)
- * @param offset     LIMIT 시작 위치 (page * size)
+ * @param offset     LIMIT 시작 위치 (page * size) — 큰 page 에서 int overflow 를 피하려 long
  * @param limit      페이지 크기
  */
 public record CompanyFileCriteria(
@@ -15,7 +15,7 @@ public record CompanyFileCriteria(
         String keyword,
         Long projectId,
         String extension,
-        int offset,
+        long offset,
         int limit
 ) {
 }
